requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
google-api-core==1.26.1
google-api-python-client==1.12.8
google-auth==1.27.1
google-auth-httplib2==0.1.0
Google-Images-Search==1.3.6
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
numpy==1.20.1
opencv-python==4.5.1.48
packaging==20.9
Pillow==8.0.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
pyparsing==2.4.7
python-resize-image==1.1.19
pytz==2021.1
requests==2.25.1
rsa==4.7.2
scipy==1.6.1
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.26.3
yapf==0.31.0
django-heroku